+++ internal/repository/question_subcategory/pg/list_subcategories_by_question_id.go
@@ -8,8 +8,8 @@ import (
 	modelRepo "github.com/Kosfedev/learn_go/internal/repository/question_subcategory/pg/model"
 )
 
-func (r *repo) ListSubcategoriesByQuestionID(ctx context.Context, questionID int) ([]int, error) {
-	idRepo := int32(questionID)
+func (r *repo) ListSubcategoriesByQuestionID(ctx context.Context, questionID int64) ([]int64, error) {
+	idRepo := questionID
 	query := db.Query{
 		Name:     "question_subcategory_repository.list_subcategories_by_question_id",
 		QueryRaw: `SELECT * FROM question_subcategory WHERE question_id = $1`,
