+++ internal/repository/questionset/pg/model/question_set_question.go
@@ -1,11 +1,11 @@
 package model
 
 type QuestionSetQuestion struct {
-	QuestionID    int32
-	QuestionSetID int32
+	QuestionID    int64
+	QuestionSetID int64
 }
 
 type NewQuestionSetQuestion struct {
-	QuestionID    int32
-	QuestionSetID int32
+	QuestionID    int64
+	QuestionSetID int64
 }
