+++ internal/service/category/create.go
@@ -6,7 +6,7 @@ import (
 	"github.com/Kosfedev/learn_go/internal/model"
 )
 
-func (qcs *serv) Create(ctx context.Context, newCategory *model.NewCategory) (int, error) {
+func (qcs *serv) Create(ctx context.Context, newCategory *model.NewCategory) (int64, error) {
 	id, err := qcs.repo.Create(ctx, newCategory)
 	if err != nil {
 		return 0, err
