+++ migrations/20250728112930_create_question_option_table.sql
@@ -2,8 +2,8 @@
 -- +goose StatementBegin
 SELECT 'up SQL query';
 CREATE TABLE question_option(
-  id SERIAL PRIMARY KEY,
-  question_id INTEGER NOT NULL REFERENCES question(id) ON DELETE CASCADE,
+  id BIGSERIAL PRIMARY KEY,
+  question_id BIGINT NOT NULL REFERENCES question(id) ON DELETE CASCADE,
   text text NOT NULL,
   is_correct boolean NOT NULL
 );
